Backend/deploymentBackend/requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions
regex==2020.7.14